# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shieldoo Gate REST API
  description: |
    Administration and observability API for Shieldoo Gate — an open-source
    supply chain security proxy for Docker, PyPI, npm, NuGet, and more.
  version: 1.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development (admin port)
...
paths:
...
  /api/v1/api-keys/{id}:
    delete:
      summary: Revoke an API key
      operationId: revokeAPIKey
      tags: [api-keys]
      description: |
        Permanently disables an API key (sets enabled=false).
        Revocation is permanent — the key cannot be re-enabled.
        Only available when both `auth.enabled` and `proxy_auth.enabled` are true.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Key revoked
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cannot revoke another user's key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Key not found
...
  /api/v1/ai/anomalies/{id}/acknowledge:
    post:
      summary: Acknowledge an anomaly for the current user
      operationId: acknowledgeAnomaly
      tags: [ai]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        "204": { description: Acknowledged }
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...